OP-327: Move strategy services to renderer, add entity to media content element
