+++ src/Renderer/ContentElementRendererStrategy.php
@@ -8,7 +8,7 @@
 
 declare(strict_types=1);
 
-namespace BitBag\SyliusCmsPlugin\Strategy;
+namespace BitBag\SyliusCmsPlugin\Renderer;
 
 use BitBag\SyliusCmsPlugin\Entity\BlockInterface;
 use BitBag\SyliusCmsPlugin\Entity\PageInterface;
