+++ src/Resources/config/services/renderer.xml
@@ -4,6 +4,11 @@
     <services>
         <defaults public="true" />
 
+        <service id="bitbag_sylius_cms_plugin.content_element_renderer_strategy" class="BitBag\SyliusCmsPlugin\Renderer\ContentElementRendererStrategy">
+            <argument type="service" id="bitbag_sylius_cms_plugin.twig.parser.content" />
+            <argument type="tagged_iterator" tag="bitbag_sylius_cms_plugin.renderer.content_element" />
+        </service>
+
         <service id="bitbag_sylius_cms_plugin.renderer.content_element.textarea" class="BitBag\SyliusCmsPlugin\Renderer\ContentElement\TextareaContentElementRenderer">
             <argument type="service" id="twig" />
             <tag name="bitbag_sylius_cms_plugin.renderer.content_element" />
@@ -12,12 +17,14 @@
         <service id="bitbag_sylius_cms_plugin.renderer.content_element.single_media" class="BitBag\SyliusCmsPlugin\Renderer\ContentElement\SingleMediaContentElementRenderer">
             <argument type="service" id="twig" />
             <argument type="service" id="bitbag_sylius_cms_plugin.twig.runtime.media" />
+            <argument type="service" id="bitbag_sylius_cms_plugin.repository.media" />
             <tag name="bitbag_sylius_cms_plugin.renderer.content_element" />
         </service>
 
         <service id="bitbag_sylius_cms_plugin.renderer.content_element.multiple_media" class="BitBag\SyliusCmsPlugin\Renderer\ContentElement\MultipleMediaContentElementRenderer">
             <argument type="service" id="twig" />
             <argument type="service" id="bitbag_sylius_cms_plugin.twig.runtime.media" />
+            <argument type="service" id="bitbag_sylius_cms_plugin.repository.media" />
             <tag name="bitbag_sylius_cms_plugin.renderer.content_element" />
         </service>
