--- src/Resources/config/services/strategy.xml
@@ -1,12 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<container xmlns="http://symfony.com/schema/dic/services" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://symfony.com/schema/dic/services http://symfony.com/schema/dic/services/services-1.0.xsd">
-    <services>
-        <defaults public="true" />
-
-        <service id="bitbag_sylius_cms_plugin.content_element_renderer_strategy" class="BitBag\SyliusCmsPlugin\Strategy\ContentElementRendererStrategy">
-            <argument type="service" id="bitbag_sylius_cms_plugin.twig.parser.content" />
-            <argument type="tagged_iterator" tag="bitbag_sylius_cms_plugin.renderer.content_element" />
-        </service>
-    </services>
-</container>
