+++ src/Resources/views/Shop/ContentElement/_multiple_media.html.twig
@@ -1,5 +1,5 @@
 <div class="{{ content_element_base_class }}__multiple-media">
     {% for element in media %}
-        {{ element }}
+        {{ element.renderedContent }}
     {% endfor %}
 </div>
