+++ src/Resources/views/Shop/ContentElement/_single_media.html.twig
@@ -1,3 +1,3 @@
 <div class="{{ content_element_base_class }}__single-media">
-    {{ media }}
+    {{ media.renderedContent }}
 </div>
